Move cluster managers into sub-module [test-yarn]
Reorganizes the cluster managers into a sub-module
``dask_gateway_server.managers``.
